LIBITD-2455. Refactored HTTP querying into HTTP/Alma gateways
Moved Alma-specific HTTP retrieval code out of "core" into "textbook",
and refactored to a more fluent ("retrieveBibs") method name. Added
generic "HttpGateway" to make a generic HTTP retrieval mechanism,
with error handling.

Updated MockAlmaServerGateway with new AlmaServerGateway interface.

https://umd-dit.atlassian.net/browse/LIBITD-2455
